Add min_points and multipler option to forecast_time_period_epiestim I add an option to etiher input min_points or a multiplier to define the minimun number of data points (size of first window) to produce an accurate fit (multiplier*serial_interval).
Because serial interval defaults was defined in the inner functions, and the windows are defined in the outer function, I created an internal helper in utils.R defining all the default values. Then, I called this helper in both outer loop (that defines min_points and windows) and inner loop (that fits data).

Readme.md passed devtools::build_readme()

Fixes #38